requirements.txt
asgiref==3.8.1
AWSIoTPythonSDK==1.5.4
bidict==0.23.1
blinker==1.9.0
boto3==1.35.92
botocore==1.35.92
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
Flask==3.1.0
Flask-SocketIO==5.5.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
msgpack==1.1.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
s3transfer==0.10.4
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
